FInished adding rate endpoint
